feat: load attributes from CMS
