+++ extension/src/prototypes/view/selection/LayerContent.tsx
@@ -157,6 +157,11 @@ export function LayerContent<T extends SupportedLayerType>({
 
   const layerName = layerTypeNames[type] ?? rootLayer?.layerName;
 
+  const plateauSpecMajorVersion =
+    rootLayerConfig?.rawDataset.__typename === "PlateauDataset"
+      ? rootLayerConfig.rawDataset.plateauSpecMinor.majorVersion
+      : 0;
+
   return (
     <>
       <List disablePadding>
@@ -238,6 +243,7 @@ export function LayerContent<T extends SupportedLayerType>({
           state={buildingSearchPanelState}
           layer={layer}
           layerId={layerId}
+          plateauSpecMajorVersion={plateauSpecMajorVersion}
         />
       )}
     </>
