+++ extension/src/shared/graphql/base/catalog/__gen__/gql.ts
@@ -13,13 +13,14 @@ import type { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-
  * Therefore it is highly recommended to use the babel or swc plugin for production.
  */
 const documents = {
-    "\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n    }\n  }\n": types.DatasetFragmentFragmentDoc,
+    "\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n      plateauSpecMinor {\n        majorVersion\n      }\n    }\n  }\n": types.DatasetFragmentFragmentDoc,
     "\nquery Areas($input: AreasInput!) {\n  areas(input: $input) {\n    id\n    code\n    name\n  }\n}\n": types.AreasDocument,
     "\nquery AreaDatasets($code: AreaCode!, $input: DatasetsInput!) {\n  area(code: $code) {\n    id\n    code\n    name\n    datasets(input: $input) {\n      ...DatasetFragment\n    }\n  }\n}\n": types.AreaDatasetsDocument,
     "\nquery Datasets($input: DatasetsInput!) {\n  datasets(input: $input) {\n    ...DatasetFragment\n  }\n}\n": types.DatasetsDocument,
     "\nquery DatasetById($id: ID!) {\n  node(id: $id) {\n    ... on Dataset {\n      ...DatasetFragment\n    }\n  }\n}\n": types.DatasetByIdDocument,
     "\nquery DatasetsByIds($ids: [ID!]!) {\n  nodes(ids: $ids) {\n    ... on Dataset {\n      ...DatasetFragment\n    }\n  }\n}\n": types.DatasetsByIdsDocument,
     "\nquery DatasetTypes($input: DatasetTypesInput) {\n  datasetTypes(input: $input) {\n    id\n    name\n    code\n    order\n  }\n}\n": types.DatasetTypesDocument,
+    "\nquery PlateauSpecs {\n  plateauSpecs {\n    majorVersion\n  }\n}\n": types.PlateauSpecsDocument,
 };
 
 /**
@@ -39,7 +40,7 @@ export function gql(source: string): unknown;
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function gql(source: "\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n    }\n  }\n"): (typeof documents)["\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n    }\n  }\n"];
+export function gql(source: "\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n      plateauSpecMinor {\n        majorVersion\n      }\n    }\n  }\n"): (typeof documents)["\n  fragment DatasetFragment on Dataset {\n    id\n    name\n    description\n    year\n    groups\n    openDataUrl\n    prefectureId\n    prefectureCode\n    cityId\n    cityCode\n    wardId\n    wardCode\n    prefecture {\n      name\n      code\n    }\n    city {\n      name\n      code\n    }\n    ward {\n      name\n      code\n    }\n    type {\n      id\n      code\n      name\n      category\n      order\n    }\n    items {\n      id\n      format\n      name\n      url\n      layers\n    }\n    admin\n    ... on PlateauDataset {\n      subname\n      plateauSpecMinor {\n        majorVersion\n      }\n    }\n  }\n"];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -64,6 +65,10 @@ export function gql(source: "\nquery DatasetsByIds($ids: [ID!]!) {\n  nodes(ids:
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function gql(source: "\nquery DatasetTypes($input: DatasetTypesInput) {\n  datasetTypes(input: $input) {\n    id\n    name\n    code\n    order\n  }\n}\n"): (typeof documents)["\nquery DatasetTypes($input: DatasetTypesInput) {\n  datasetTypes(input: $input) {\n    id\n    name\n    code\n    order\n  }\n}\n"];
+/**
+ * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function gql(source: "\nquery PlateauSpecs {\n  plateauSpecs {\n    majorVersion\n  }\n}\n"): (typeof documents)["\nquery PlateauSpecs {\n  plateauSpecs {\n    majorVersion\n  }\n}\n"];
 
 export function gql(source: string) {
   return (documents as any)[source] ?? {};
