+++ extension/src/shared/graphql/base/catalog/__gen__/graphql.ts
@@ -172,6 +172,8 @@ export type CityGmlDataset = Node & {
 export type Dataset = {
   /** 管理者用 */
   admin?: Maybe<Scalars['Any']['output']>;
+  /** PLATEAU ARで閲覧可能なデータセットかどうか。 */
+  ar: Scalars['Boolean']['output'];
   /** データセットが属する市。 */
   city?: Maybe<City>;
   /** データセットが属する市コード。先頭に都道府県コードを含む5桁の数字から成る文字列です。 */
@@ -300,6 +302,11 @@ export type DatasetTypesInput = {
 
 /** データセットを検索するためのクエリ。 */
 export type DatasetsInput = {
+  /**
+   * PLATEAU ARで閲覧可能なデータセットを含めるかどうか。
+   * trueの場合はARで閲覧可能なデータセットのみ、falseの場合はARで閲覧不可能なデータセットのみを返します。
+   */
+  ar?: InputMaybe<Scalars['Boolean']['input']>;
   /** データセットの地域コード（都道府県コードや市区町村コードが使用可能）。複数指定するとOR条件で検索を行います。 */
   areaCodes?: InputMaybe<Array<Scalars['AreaCode']['input']>>;
   /** 検索結果から除外するデータセットの種類コード。種類コードは例えば "bldg"（建築物モデル）の他、"plateau"（PLATEAU都市モデルデータセット）、"related"（関連データセット）、"generic"（その他のデータセット）が使用可能です。 */
@@ -337,6 +344,8 @@ export type GenericDataset = Dataset & Node & {
   __typename?: 'GenericDataset';
   /** 管理者用 */
   admin?: Maybe<Scalars['Any']['output']>;
+  /** PLATEAU ARで閲覧可能なデータセットかどうか。 */
+  ar: Scalars['Boolean']['output'];
   /** データセットが属する市。 */
   city?: Maybe<City>;
   /** データセットが属する市コード。先頭に都道府県コードを含む5桁の数字から成る文字列です。 */
@@ -432,6 +441,8 @@ export type PlateauDataset = Dataset & Node & {
   __typename?: 'PlateauDataset';
   /** 管理者用 */
   admin?: Maybe<Scalars['Any']['output']>;
+  /** PLATEAU ARで閲覧可能なデータセットかどうか。 */
+  ar: Scalars['Boolean']['output'];
   /** データセットが属する市。 */
   city?: Maybe<City>;
   /** データセットが属する市コード。先頭に都道府県コードを含む5桁の数字から成る文字列です。 */
@@ -488,9 +499,9 @@ export type PlateauDataset = Dataset & Node & {
 /** PLATEAU都市モデルのデータセットのアイテム。 */
 export type PlateauDatasetItem = DatasetItem & Node & {
   __typename?: 'PlateauDatasetItem';
-  /** 浸水規模。地物型が洪水・高潮・津波・内水浸水想定区域モデル（fld・htd・tnm・ifld）の場合のみ存在します。 */
+  /** 浸水規模。地物型が災害リスク（浸水）モデルの場合のみ存在することがあります。 */
   floodingScale?: Maybe<FloodingScale>;
-  /** 浸水規模の枝番。地物型が洪水・高潮・津波・内水浸水想定区域モデル（fld・htd・tnm・ifld）の場合のみ存在することがあります。 */
+  /** 浸水規模の枝番。地物型が災害リスク（浸水）モデルの場合のみ存在することがあります。 */
   floodingScaleSuffix?: Maybe<Scalars['String']['output']>;
   /** データセットのアイテムのフォーマット。 */
   format: DatasetFormat;
@@ -502,6 +513,8 @@ export type PlateauDatasetItem = DatasetItem & Node & {
   layers?: Maybe<Array<Scalars['String']['output']>>;
   /** データセットのアイテムのLOD（詳細度・Level of Detail）。1、2、3、4などの整数値です。 */
   lod?: Maybe<Scalars['Int']['output']>;
+  /** データセットのアイテムのLOD（詳細度・Level of Detail）のうち、小数点以下の値が存在する場合に定義されます。例えばLOD3.1の場合は1、3.0の場合は0となります。LODがnullの場合はnullとなります。 */
+  lodEx?: Maybe<Scalars['Int']['output']>;
   /** データセットのアイテム名。 */
   name: Scalars['String']['output'];
   /** データセットのアイテムが属するデータセット。 */
@@ -523,7 +536,7 @@ export type PlateauDatasetType = DatasetType & Node & {
   code: Scalars['String']['output'];
   /** データセット（DatasetInput内のincludeTypesとexcludeTypesの指定は無視されます）。 */
   datasets: Array<PlateauDataset>;
-  /** 洪水・高潮・津波・内水浸水想定区域モデルを表す種類かどうか。河川などの情報が利用可能です。 */
+  /** 災害リスク（浸水）モデルかどうか。河川などの情報が利用可能です。 */
   flood: Scalars['Boolean']['output'];
   id: Scalars['ID']['output'];
   /** データセットの種類名。 */
@@ -677,6 +690,8 @@ export type RelatedDataset = Dataset & Node & {
   __typename?: 'RelatedDataset';
   /** 管理者用 */
   admin?: Maybe<Scalars['Any']['output']>;
+  /** PLATEAU ARで閲覧可能なデータセットかどうか。 */
+  ar: Scalars['Boolean']['output'];
   /** データセットが属する市。 */
   city?: Maybe<City>;
   /** データセットが属する市コード。先頭に都道府県コードを含む5桁の数字から成る文字列です。 */
@@ -836,7 +851,7 @@ export type WardDatasetsArgs = {
 
 type DatasetFragment_GenericDataset_Fragment = { __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> };
 
-type DatasetFragment_PlateauDataset_Fragment = { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> };
+type DatasetFragment_PlateauDataset_Fragment = { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> };
 
 type DatasetFragment_RelatedDataset_Fragment = { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> };
 
@@ -855,28 +870,28 @@ export type AreaDatasetsQueryVariables = Exact<{
 }>;
 
 
-export type AreaDatasetsQuery = { __typename?: 'Query', area?: { __typename?: 'City', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | { __typename?: 'Prefecture', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | { __typename?: 'Ward', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | null };
+export type AreaDatasetsQuery = { __typename?: 'Query', area?: { __typename?: 'City', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | { __typename?: 'Prefecture', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | { __typename?: 'Ward', id: string, code: any, name: string, datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> } | null };
 
 export type DatasetsQueryVariables = Exact<{
   input: DatasetsInput;
 }>;
 
 
-export type DatasetsQuery = { __typename?: 'Query', datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> };
+export type DatasetsQuery = { __typename?: 'Query', datasets: Array<{ __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> }> };
 
 export type DatasetByIdQueryVariables = Exact<{
   id: Scalars['ID']['input'];
 }>;
 
 
-export type DatasetByIdQuery = { __typename?: 'Query', node?: { __typename?: 'City' } | { __typename?: 'CityGMLDataset' } | { __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'GenericDatasetItem' } | { __typename?: 'GenericDatasetType' } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDatasetItem' } | { __typename?: 'PlateauDatasetType' } | { __typename?: 'PlateauSpec' } | { __typename?: 'PlateauSpecMinor' } | { __typename?: 'Prefecture' } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDatasetItem' } | { __typename?: 'RelatedDatasetType' } | { __typename?: 'Ward' } | null };
+export type DatasetByIdQuery = { __typename?: 'Query', node?: { __typename?: 'City' } | { __typename?: 'CityGMLDataset' } | { __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'GenericDatasetItem' } | { __typename?: 'GenericDatasetType' } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDatasetItem' } | { __typename?: 'PlateauDatasetType' } | { __typename?: 'PlateauSpec' } | { __typename?: 'PlateauSpecMinor' } | { __typename?: 'Prefecture' } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDatasetItem' } | { __typename?: 'RelatedDatasetType' } | { __typename?: 'Ward' } | null };
 
 export type DatasetsByIdsQueryVariables = Exact<{
   ids: Array<Scalars['ID']['input']> | Scalars['ID']['input'];
 }>;
 
 
-export type DatasetsByIdsQuery = { __typename?: 'Query', nodes: Array<{ __typename?: 'City' } | { __typename?: 'CityGMLDataset' } | { __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'GenericDatasetItem' } | { __typename?: 'GenericDatasetType' } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDatasetItem' } | { __typename?: 'PlateauDatasetType' } | { __typename?: 'PlateauSpec' } | { __typename?: 'PlateauSpecMinor' } | { __typename?: 'Prefecture' } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDatasetItem' } | { __typename?: 'RelatedDatasetType' } | { __typename?: 'Ward' } | null> };
+export type DatasetsByIdsQuery = { __typename?: 'Query', nodes: Array<{ __typename?: 'City' } | { __typename?: 'CityGMLDataset' } | { __typename?: 'GenericDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'GenericDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'GenericDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'GenericDatasetItem' } | { __typename?: 'GenericDatasetType' } | { __typename?: 'PlateauDataset', subname?: string | null, id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, plateauSpecMinor: { __typename?: 'PlateauSpecMinor', majorVersion: number }, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'PlateauDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'PlateauDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'PlateauDatasetItem' } | { __typename?: 'PlateauDatasetType' } | { __typename?: 'PlateauSpec' } | { __typename?: 'PlateauSpecMinor' } | { __typename?: 'Prefecture' } | { __typename?: 'RelatedDataset', id: string, name: string, description?: string | null, year: number, groups?: Array<string> | null, openDataUrl?: string | null, prefectureId?: string | null, prefectureCode?: any | null, cityId?: string | null, cityCode?: any | null, wardId?: string | null, wardCode?: any | null, admin?: any | null, prefecture?: { __typename?: 'Prefecture', name: string, code: any } | null, city?: { __typename?: 'City', name: string, code: any } | null, ward?: { __typename?: 'Ward', name: string, code: any } | null, type: { __typename?: 'RelatedDatasetType', id: string, code: string, name: string, category: DatasetTypeCategory, order: number }, items: Array<{ __typename?: 'RelatedDatasetItem', id: string, format: DatasetFormat, name: string, url: string, layers?: Array<string> | null }> } | { __typename?: 'RelatedDatasetItem' } | { __typename?: 'RelatedDatasetType' } | { __typename?: 'Ward' } | null> };
 
 export type DatasetTypesQueryVariables = Exact<{
   input?: InputMaybe<DatasetTypesInput>;
@@ -885,10 +900,16 @@ export type DatasetTypesQueryVariables = Exact<{
 
 export type DatasetTypesQuery = { __typename?: 'Query', datasetTypes: Array<{ __typename?: 'GenericDatasetType', id: string, name: string, code: string, order: number } | { __typename?: 'PlateauDatasetType', id: string, name: string, code: string, order: number } | { __typename?: 'RelatedDatasetType', id: string, name: string, code: string, order: number }> };
 
-export const DatasetFragmentFragmentDoc = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}}]}}]}}]} as unknown as DocumentNode<DatasetFragmentFragment, unknown>;
+export type PlateauSpecsQueryVariables = Exact<{ [key: string]: never; }>;
+
+
+export type PlateauSpecsQuery = { __typename?: 'Query', plateauSpecs: Array<{ __typename?: 'PlateauSpec', majorVersion: number }> };
+
+export const DatasetFragmentFragmentDoc = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}},{"kind":"Field","name":{"kind":"Name","value":"plateauSpecMinor"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]}}]} as unknown as DocumentNode<DatasetFragmentFragment, unknown>;
 export const AreasDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Areas"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"AreasInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"areas"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}}]}}]}}]} as unknown as DocumentNode<AreasQuery, AreasQueryVariables>;
-export const AreaDatasetsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"AreaDatasets"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"code"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"AreaCode"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"DatasetsInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"area"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"code"},"value":{"kind":"Variable","name":{"kind":"Name","value":"code"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"datasets"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}}]}}]}}]} as unknown as DocumentNode<AreaDatasetsQuery, AreaDatasetsQueryVariables>;
-export const DatasetsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Datasets"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"DatasetsInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"datasets"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}}]}}]}}]} as unknown as DocumentNode<DatasetsQuery, DatasetsQueryVariables>;
-export const DatasetByIdDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DatasetById"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"node"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}}]}}]}}]} as unknown as DocumentNode<DatasetByIdQuery, DatasetByIdQueryVariables>;
-export const DatasetsByIdsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DatasetsByIds"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"ids"}},"type":{"kind":"NonNullType","type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"nodes"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"ids"},"value":{"kind":"Variable","name":{"kind":"Name","value":"ids"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}}]}}]}}]} as unknown as DocumentNode<DatasetsByIdsQuery, DatasetsByIdsQueryVariables>;
+export const AreaDatasetsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"AreaDatasets"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"code"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"AreaCode"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"DatasetsInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"area"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"code"},"value":{"kind":"Variable","name":{"kind":"Name","value":"code"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"datasets"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}},{"kind":"Field","name":{"kind":"Name","value":"plateauSpecMinor"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]}}]} as unknown as DocumentNode<AreaDatasetsQuery, AreaDatasetsQueryVariables>;
+export const DatasetsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Datasets"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"DatasetsInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"datasets"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}},{"kind":"Field","name":{"kind":"Name","value":"plateauSpecMinor"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]}}]} as unknown as DocumentNode<DatasetsQuery, DatasetsQueryVariables>;
+export const DatasetByIdDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DatasetById"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"node"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}},{"kind":"Field","name":{"kind":"Name","value":"plateauSpecMinor"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]}}]} as unknown as DocumentNode<DatasetByIdQuery, DatasetByIdQueryVariables>;
+export const DatasetsByIdsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DatasetsByIds"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"ids"}},"type":{"kind":"NonNullType","type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"nodes"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"ids"},"value":{"kind":"Variable","name":{"kind":"Name","value":"ids"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"DatasetFragment"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"DatasetFragment"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Dataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"year"}},{"kind":"Field","name":{"kind":"Name","value":"groups"}},{"kind":"Field","name":{"kind":"Name","value":"openDataUrl"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureId"}},{"kind":"Field","name":{"kind":"Name","value":"prefectureCode"}},{"kind":"Field","name":{"kind":"Name","value":"cityId"}},{"kind":"Field","name":{"kind":"Name","value":"cityCode"}},{"kind":"Field","name":{"kind":"Name","value":"wardId"}},{"kind":"Field","name":{"kind":"Name","value":"wardCode"}},{"kind":"Field","name":{"kind":"Name","value":"prefecture"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"city"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"ward"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}}]}},{"kind":"Field","name":{"kind":"Name","value":"type"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"category"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"format"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"layers"}}]}},{"kind":"Field","name":{"kind":"Name","value":"admin"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"PlateauDataset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"subname"}},{"kind":"Field","name":{"kind":"Name","value":"plateauSpecMinor"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]}}]} as unknown as DocumentNode<DatasetsByIdsQuery, DatasetsByIdsQueryVariables>;
 export const DatasetTypesDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DatasetTypes"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DatasetTypesInput"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"datasetTypes"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"code"}},{"kind":"Field","name":{"kind":"Name","value":"order"}}]}}]}}]} as unknown as DocumentNode<DatasetTypesQuery, DatasetTypesQueryVariables>;
+export const PlateauSpecsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"PlateauSpecs"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"plateauSpecs"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"majorVersion"}}]}}]}}]} as unknown as DocumentNode<PlateauSpecsQuery, PlateauSpecsQueryVariables>;
