+++ extension/src/shared/graphql/base/catalog/queries/plateauSpecs.ts
@@ -1,0 +1,9 @@
+import { gql } from "../__gen__/gql";
+
+export const PLATEAU_SPECS = gql(`
+query PlateauSpecs {
+  plateauSpecs {
+    majorVersion
+  }
+}
+`);
