+++ extension/src/shared/graphql/hooks/catalog/index.ts
@@ -1,3 +1,4 @@
 export * from "./dataset";
 export * from "./datasetType";
 export * from "./area";
+export * from "./plateauSpecs";
