+++ extension/src/shared/plateau/featureInspector/index.ts
@@ -1,3 +1,4 @@
 export { getAttributeLabel, getRootFields } from "./attributes";
 export { replaceUnknownDate } from "./clearfyDate";
+export { loadAttributes } from "./loadAttributes";
 export * from "./utils";
