+++ extension/src/shared/view-layers/plateau-3dtiles/BuildingLayer.tsx
@@ -22,6 +22,7 @@ export interface BuildingLayerModelParams extends LayerModelParams, PlateauTiles
   municipalityCode: string;
   title: string;
   textured?: boolean;
+  version: number;
 }
 
 export interface BuildingLayerModel extends LayerModel, PlateauTilesetLayerState {
@@ -29,6 +30,7 @@ export interface BuildingLayerModel extends LayerModel, PlateauTilesetLayerState
   title: string;
   showWireframeAtom: PrimitiveAtom<boolean>;
   textured: boolean;
+  version: number;
 }
 
 export function createBuildingLayer(
@@ -42,6 +44,7 @@ export function createBuildingLayer(
     municipalityCode: params.municipalityCode,
     title: params.title,
     showWireframeAtom: atom(false),
+    version: params.version,
   };
 }
 
@@ -65,6 +68,7 @@ export const BuildingLayer: FC<LayerProps<typeof BUILDING_LAYER>> = ({
   colorMapAtom,
   colorRangeAtom,
   componentAtoms,
+  version,
   // showWireframeAtom,
 }) => {
   const hidden = useAtomValue(hiddenAtom);
@@ -117,6 +121,7 @@ export const BuildingLayer: FC<LayerProps<typeof BUILDING_LAYER>> = ({
         colorRangeAtom={colorRangeAtom}
         searchedFeaturesAtom={searchedFeaturesAtom}
         selections={selections as ScreenSpaceSelectionEntry<typeof TILESET_FEATURE>[]}
+        version={version}
         // showWireframe={showWireframe}
 
         // Field components
