+++ extension/src/toolbar/Widget.tsx
@@ -42,6 +42,7 @@ type DefaultProps = {
   plateauAccessToken?: string;
   catalogURL?: string;
   catalogURLForAdmin?: string;
+  datasetAttributesURL?: string;
   projectName?: string;
   googleStreetViewAPIKey?: string;
   geojsonURL?: string;
@@ -80,6 +81,7 @@ export const Widget: FC<Props> = memo(function WidgetPresenter({ widget, inEdito
         plateauToken={widget.property.default.plateauAccessToken}
         catalogUrl={widget.property.default.catalogURL}
         catalogURLForAdmin={widget.property.default.catalogURLForAdmin}
+        datasetAttributesURL={widget.property.default.datasetAttributesURL}
         geoUrl={widget.property.default.geoURL}
         gsiTileURL={widget.property.default.gsiTileURL}
         googleStreetViewAPIKey={
